+++ src/main/java/com/xiaoleilu/hutool/crypto/digest/DigestAlgorithm.java
@@ -11,7 +11,7 @@ public enum DigestAlgorithm {
 	MD5("MD5"), 
 	SHA1("SHA-1"), 
 	SHA256("SHA-256"), 
-	SHA348("SHA-348"), 
+	SHA384("SHA-384"), 
 	SHA512("SHA-512");
 
 	private String value;